chore: move query context to a shared location
