+++ src/flows/context/flow.query.tsx
@@ -4,7 +4,7 @@ import {getVariables} from 'src/variables/selectors'
 import {FlowContext} from 'src/flows/context/flow.current'
 import {RunModeContext, RunMode} from 'src/flows/context/runMode'
 import {ResultsContext} from 'src/flows/context/results'
-import {QueryContext, simplify} from 'src/flows/context/query'
+import {QueryContext, simplify} from 'src/shared/contexts/query'
 import {event} from 'src/cloud/utils/reporting'
 import {FluxResult, QueryScope} from 'src/types/flows'
 import {PIPE_DEFINITIONS, PROJECT_NAME} from 'src/flows'
