+++ src/writeData/components/ClientCodeQueryHelper.tsx
@@ -4,7 +4,7 @@ import {format_from_js_file} from '@influxdata/flux'
 
 // Utils
 import {parse} from 'src/external/parser'
-import {find} from 'src/flows/context/query'
+import {find} from 'src/shared/contexts/query'
 import {WriteDataDetailsContext} from 'src/writeData/components/WriteDataDetailsContext'
 
 // Constants
